requirements.txt
uvicorn==0.32.1
pydantic[email]==2.10.3
pydantic==2.10.3
pydantic_core==2.27.1
python-multipart==0.0.19
fastapi==0.115.5
amocrm-api==2.6.1